Allow scoped runtime sandbox paths
