+++ crates/csa-resource/src/isolation_plan.rs
@@ -1,9 +1,7 @@
 //! Isolation plan: combines resource and filesystem capabilities into a
 //! single, builder-configured plan that executors can apply uniformly.
 use std::collections::HashMap;
-use std::path::{Component, Path, PathBuf};
-
-use anyhow::Context;
+use std::path::{Path, PathBuf};
 
 use crate::filesystem_sandbox::FilesystemCapability;
 use crate::sandbox::ResourceCapability;
@@ -12,6 +10,17 @@ pub const DEFAULT_SANDBOX_TMPDIR: &str = "/tmp";
 
 #[path = "isolation_plan_codex.rs"]
 mod codex_paths;
+#[path = "isolation_plan_runtime_path.rs"]
+mod runtime_path;
+#[path = "isolation_plan_validation.rs"]
+mod validation;
+#[cfg(test)]
+use runtime_path::is_xdg_runtime_child_path;
+use runtime_path::{detect_superproject_root, home_dir, is_sensitive_system_path};
+pub use validation::{
+    canonicalize_through_existing_ancestors, resolve_writable_paths, validate_readable_paths,
+    validate_writable_paths,
+};
 
 /// Sandbox enforcement mode.
 ///
@@ -422,274 +431,6 @@ fn sandbox_tmpdir_for_capability(filesystem: FilesystemCapability, session_dir: 
     }
 }
 
-/// Strictly validate writable sandbox paths against default safe roots.
-/// # Errors
-///
-/// Returns an error for root, sensitive system paths, or paths outside
-/// `project_root`, the user home directory, and `/tmp`.
-pub fn validate_writable_paths(paths: &[PathBuf], project_root: &Path) -> anyhow::Result<()> {
-    resolve_writable_paths_impl(paths, project_root, false).map(|_| ())
-}
-
-pub fn resolve_writable_paths(
-    paths: &[PathBuf],
-    project_root: &Path,
-) -> anyhow::Result<Vec<PathBuf>> {
-    resolve_writable_paths_impl(paths, project_root, true)
-}
-
-fn resolve_writable_paths_impl(
-    paths: &[PathBuf],
-    project_root: &Path,
-    allow_outside_default_roots: bool,
-) -> anyhow::Result<Vec<PathBuf>> {
-    validate_sandbox_paths(
-        paths,
-        project_root,
-        PathValidationOptions {
-            kind: "writable_paths",
-            require_absolute: false,
-            require_exists: false,
-            reject_tmp_root: false,
-            canonicalize_for_allowlist: true,
-            allow_requested_path_for_allowlist: true,
-            allow_outside_default_roots,
-        },
-    )
-}
-
-/// Validate that readable paths are safe to expose into the sandbox.
-///
-/// Read-only binds are stricter than writable paths: every path must be
-/// absolute, must exist on disk, `/tmp` itself is forbidden, and symlinked
-/// paths are validated against the canonical target to prevent bind-mounting a
-/// safe-looking path that resolves somewhere outside the allowlist.
-pub fn validate_readable_paths(paths: &[PathBuf], project_root: &Path) -> anyhow::Result<()> {
-    validate_sandbox_paths(
-        paths,
-        project_root,
-        PathValidationOptions {
-            kind: "readable_paths",
-            require_absolute: true,
-            require_exists: true,
-            reject_tmp_root: true,
-            canonicalize_for_allowlist: true,
-            allow_requested_path_for_allowlist: false,
-            allow_outside_default_roots: false,
-        },
-    )
-    .map(|_| ())
-}
-
-/// Canonicalize `path` through its deepest existing ancestor.
-/// Missing tail components are re-attached, allowing writable directories that
-/// may be pre-created later via `create_dir_all()`.
-pub fn canonicalize_through_existing_ancestors(path: &Path) -> anyhow::Result<PathBuf> {
-    let mut candidate = path.to_path_buf();
-    let mut missing_suffix = Vec::new();
-
-    loop {
-        if candidate.as_os_str().is_empty() {
-            let mut resolved = std::env::current_dir().with_context(|| {
-                format!(
-                    "failed to resolve current directory while canonicalizing {}",
-                    path.display()
-                )
-            })?;
-            for component in missing_suffix.iter().rev() {
-                resolved.push(component);
-            }
-            return Ok(resolved);
-        }
-
-        match candidate.canonicalize() {
-            Ok(mut resolved) => {
-                for component in missing_suffix.iter().rev() {
-                    resolved.push(component);
-                }
-                return Ok(resolved);
-            }
-            Err(error) => match candidate.try_exists() {
-                Ok(true) => {
-                    return Err(error).with_context(|| {
-                        format!(
-                            "failed to canonicalize existing path {} while resolving {}",
-                            candidate.display(),
-                            path.display()
-                        )
-                    });
-                }
-                Ok(false) => {
-                    let component = candidate.file_name().with_context(|| {
-                        format!(
-                            "path {} has no existing ancestor to canonicalize through",
-                            path.display()
-                        )
-                    })?;
-                    missing_suffix.push(component.to_os_string());
-                    candidate.pop();
-                }
-                Err(exists_error) => {
-                    return Err(exists_error).with_context(|| {
-                        format!(
-                            "failed to probe path existence while resolving {}",
-                            path.display()
-                        )
-                    });
-                }
-            },
-        }
-    }
-}
-
-struct PathValidationOptions<'a> {
-    kind: &'a str,
-    require_absolute: bool,
-    require_exists: bool,
-    reject_tmp_root: bool,
-    canonicalize_for_allowlist: bool,
-    allow_requested_path_for_allowlist: bool,
-    allow_outside_default_roots: bool,
-}
-
-fn validate_sandbox_paths(
-    paths: &[PathBuf],
-    project_root: &Path,
-    options: PathValidationOptions<'_>,
-) -> anyhow::Result<Vec<PathBuf>> {
-    let home = home_dir().unwrap_or_else(|| PathBuf::from("/nonexistent"));
-    let project_root = canonicalize_or_fallback(project_root);
-    let project_root_for_join = project_root.clone();
-    let home = canonicalize_or_fallback(home.as_path());
-    let tmp_root = canonicalize_or_fallback(Path::new("/tmp"));
-    let allowed_parents = [project_root, home, tmp_root];
-    let mut rejected = Vec::new();
-    let mut resolved_paths = Vec::with_capacity(paths.len());
-
-    for path in paths {
-        let validated = match validate_single_path(path, &options, project_root_for_join.as_path())
-        {
-            Ok(candidate) => candidate,
-            Err(reason) => {
-                rejected.push(format!("{} ({reason})", path.display()));
-                continue;
-            }
-        };
-
-        let is_allowed = options.allow_outside_default_roots
-            || allowed_parents
-                .iter()
-                .any(|parent| validated.resolved.starts_with(parent))
-            || (options.allow_requested_path_for_allowlist
-                && allowed_parents
-                    .iter()
-                    .any(|parent| validated.requested.starts_with(parent)));
-        if !is_allowed {
-            rejected.push(format!(
-                "{} (resolved {}; outside allowed roots: home, /tmp, project root)",
-                path.display(),
-                validated.resolved.display()
-            ));
-            continue;
-        }
-        resolved_paths.push(validated.resolved);
-    }
-
-    if rejected.is_empty() {
-        return Ok(resolved_paths);
-    }
-
-    anyhow::bail!(
-        "{} validation failed: rejected paths [{}]. Allowed: subpaths of home dir, /tmp, or project root",
-        options.kind,
-        rejected.join(", ")
-    );
-}
-
-struct ValidatedPath {
-    requested: PathBuf,
-    resolved: PathBuf,
-}
-
-fn validate_single_path(
-    path: &Path,
-    options: &PathValidationOptions<'_>,
-    project_root: &Path,
-) -> anyhow::Result<ValidatedPath> {
-    if path == Path::new("/") {
-        anyhow::bail!("root path is forbidden");
-    }
-    if options.reject_tmp_root && path == Path::new("/tmp") {
-        anyhow::bail!("/tmp itself is forbidden; expose a specific sub-path instead");
-    }
-    if options.require_absolute && !path.is_absolute() {
-        anyhow::bail!("path must be absolute");
-    }
-    let requested = normalize_path_components(if path.is_absolute() {
-        path.to_path_buf()
-    } else {
-        project_root.join(path)
-    });
-    if requested == Path::new("/") {
-        anyhow::bail!("root path is forbidden");
-    }
-    if options.reject_tmp_root && requested == Path::new("/tmp") {
-        anyhow::bail!("/tmp itself is forbidden; expose a specific sub-path instead");
-    }
-    let path_exists = !options.require_exists
-        || requested.try_exists().with_context(|| {
-            format!(
-                "failed to probe path '{}' before sandbox launch",
-                path.display()
-            )
-        })?;
-    if !path_exists {
-        anyhow::bail!(
-            "path '{}' does not exist. Create it first or remove the flag.",
-            path.display()
-        );
-    }
-
-    if !options.canonicalize_for_allowlist {
-        return Ok(ValidatedPath {
-            requested: requested.clone(),
-            resolved: requested,
-        });
-    }
-
-    let resolved = canonicalize_through_existing_ancestors(&requested)?;
-    if is_sensitive_system_path(&resolved) {
-        anyhow::bail!("resolved path {} is forbidden", resolved.display());
-    }
-    Ok(ValidatedPath {
-        requested,
-        resolved,
-    })
-}
-
-fn canonicalize_or_fallback(path: &Path) -> PathBuf {
-    path.canonicalize().unwrap_or_else(|_| path.to_path_buf())
-}
-
-fn normalize_path_components(path: PathBuf) -> PathBuf {
-    let mut normalized = PathBuf::new();
-    for component in path.components() {
-        match component {
-            Component::Prefix(_) | Component::RootDir | Component::Normal(_) => {
-                normalized.push(component.as_os_str());
-            }
-            Component::CurDir => {}
-            Component::ParentDir => {
-                if normalized.as_os_str().is_empty() || normalized == Path::new("/") {
-                    continue;
-                }
-                normalized.pop();
-            }
-        }
-    }
-    normalized
-}
-
 /// Add `dir` to `paths` if it exists, otherwise pre-create it when a
 /// non-root ancestor exists (bwrap `--bind` requires the source path to exist).
 ///
@@ -735,57 +476,6 @@ fn add_dir_or_creatable_parent(paths: &mut Vec<PathBuf>, dir: &Path) -> bool {
     } else {
         false
     }
-}
-
-/// Reject paths under sensitive system directories that should never be
-/// writable inside a sandbox.  Allows legitimate paths like home dirs,
-/// `/tmp`, `/usr/local/share/mise`, etc.
-fn is_sensitive_system_path(path: &Path) -> bool {
-    /// Prefixes that indicate sensitive system directories.
-    const SENSITIVE_PREFIXES: &[&str] = &[
-        "/etc", "/var/lib", "/var/log", "/var/run", "/boot", "/sbin", "/bin", "/lib", "/lib64",
-        "/sys", "/proc", "/dev", "/run",
-    ];
-
-    for prefix in SENSITIVE_PREFIXES {
-        if path.starts_with(prefix) {
-            return true;
-        }
-    }
-    // Reject bare root path
-    path == Path::new("/")
-}
-
-/// Portable home-directory lookup (avoids pulling in the `dirs` crate).
-fn home_dir() -> Option<PathBuf> {
-    std::env::var_os("HOME").map(PathBuf::from)
-}
-
-/// Detect whether `project_root` is inside a git submodule and return the
-/// superproject root if so.
-///
-/// A git submodule has a `.git` **file** (not directory) containing a
-/// `gitdir:` pointer.  We walk ancestors looking for the nearest directory
-/// that has a `.git` *directory* — that is the superproject root.
-///
-/// Returns `None` when `project_root` is not a submodule (`.git` is a
-/// directory or does not exist).
-fn detect_superproject_root(project_root: &Path) -> Option<PathBuf> {
-    let dot_git = project_root.join(".git");
-
-    // Only trigger when .git is a file (submodule marker).
-    if !dot_git.is_file() {
-        return None;
-    }
-
-    // Walk ancestors (skip project_root itself) looking for a .git directory.
-    for ancestor in project_root.ancestors().skip(1) {
-        if ancestor.join(".git").is_dir() {
-            return Some(ancestor.to_path_buf());
-        }
-    }
-
-    None
 }
 
 #[cfg(test)]
